Split file signing classes
